added arrayOfLists ds, clearer variable names, improved memory dump

+++ datatype_functions.c
@@ -1,7 +1,7 @@
 #include "datatype_functions.h"
 #include <stdlib.h>
 
-char *base10to16(int number) {
+char *base10to16(long number) {
 	char *hex = (char *) malloc(10 * sizeof(char));
 	int i = 0;
 
@@ -22,6 +22,7 @@ char *base10to16(int number) {
 		hex[i++] = digit;
 		number /= 16;
 	}
+
 	hex[i] = '\0';
 
 	int j = 0;
@@ -34,9 +35,10 @@ char *base10to16(int number) {
 	return hex;
 }
 
-int base16to10(char *hex) {
-	int number = 0;
+long base16to10(char *hex) {
+	long number = 0;
 	int i = 0;
+
 	while (hex[i] != '\0') {
 		number *= 16;
 		if (hex[i] >= '0' && hex[i] <= '9') {
@@ -46,6 +48,7 @@ int base16to10(char *hex) {
 		}
 		i++;
 	}
+
 	return number;
 }
 
@@ -54,10 +57,12 @@ char *addHexaPrefix(char * hex) {
 	newHex[0] = '0';
 	newHex[1] = 'x';
 	int i = 2;
+
 	while (hex[i - 2] != '\0') {
 		newHex[i] = hex[i - 2];
 		i++;
 	}
+
 	newHex[i] = '\0';
 	return newHex;
 }
@@ -65,10 +70,12 @@ char *addHexaPrefix(char * hex) {
 char *removeHexaPrefix(char * hex) {
 	char *newHex = (char *) malloc(10 * sizeof(char));
 	int i = 2;
+
 	while (hex[i] != '\0') {
 		newHex[i - 2] = hex[i];
 		i++;
 	}
+
 	newHex[i - 2] = '\0';
 	return newHex;
 }
